# manager-zedaapi/public/openapi-content-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zé da API - Content Management API",
    "description": "API completa para gerenciamento externo de conteudo da plataforma Zé da API.\n\nTodos os endpoints requerem autenticacao via header `Authorization: Bearer <CONTENT_API_KEY>`.\n\nA chave e definida na variavel de ambiente `CONTENT_API_KEY` do servidor.",
    "version": "1.0.0",
    "contact": {
      "name": "Zé da API",
      "url": "https://zedaapi.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    },
    {
      "url": "https://app.zedaapi.com",
      "description": "Producao"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CONTENT_API_KEY definida no .env do servidor. Minimo 32 caracteres."
      }
    },
    "schemas": {
...
      "GlossaryTerm": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "example": "cm5gls001"
          },
          "term": {
            "type": "string",
            "example": "Webhook"
          },
          "slug": {
            "type": "string",
            "example": "webhook"
          },
          "definition": {
            "type": "string",
            "example": "Mecanismo de notificacao HTTP que envia dados em tempo real quando um evento ocorre."
          },
          "content": {
            "type": "string",
            "nullable": true,
            "example": "<p>Webhooks sao callbacks HTTP que permitem...</p>"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ],
            "example": "published"
          },
          "seoTitle": {
            "type": "string",
            "nullable": true
          },
          "seoDescription": {
            "type": "string",
            "nullable": true
          },
          "relatedSlugs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "api",
              "http",
              "callback"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Mensagem de erro"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "API key ausente ou invalida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "missing": {
                "summary": "Header ausente",
                "value": {
                  "error": "Missing Authorization header"
                }
              },
              "invalid": {
                "summary": "Chave invalida",
                "value": {
                  "error": "Invalid API key"
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Recurso nao encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Not found"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflito (slug duplicado)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "A post with this slug already exists"
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/admin/content/glossary/terms": {
      "get": {
        "tags": [
          "Glossary Terms"
        ],
        "summary": "Listar termos do glossario",
        "description": "Retorna lista paginada de termos com filtros por busca e letra inicial.",
        "operationId": "listGlossaryTerms",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "example": 1
          },
          {
            "name": "search",
            "in": "query",
            "description": "Busca por nome do termo",
            "schema": {
              "type": "string"
            },
            "example": "webhook"
          },
          {
            "name": "letter",
            "in": "query",
            "description": "Filtrar por letra inicial (A-Z)",
            "schema": {
              "type": "string",
              "maxLength": 1
            },
            "example": "W"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de termos",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "cm5gls001",
                      "term": "Webhook",
                      "slug": "webhook",
                      "definition": "Mecanismo de notificacao HTTP que envia dados em tempo real quando um evento ocorre.",
                      "content": "<p>Webhooks sao callbacks HTTP que permitem...</p>",
                      "status": "published",
                      "relatedSlugs": [
                        "api",
                        "http",
                        "callback"
                      ]
                    },
                    {
                      "id": "cm5gls002",
                      "term": "WhatsApp Business API",
                      "slug": "whatsapp-business-api",
                      "definition": "Interface de programacao que permite empresas enviarem e receberem mensagens pelo WhatsApp.",
                      "content": null,
                      "status": "published",
                      "relatedSlugs": [
                        "webhook",
                        "api"
                      ]
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "pageSize": 20,
                    "total": 2,
                    "totalPages": 1
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Glossary Terms"
        ],
        "summary": "Criar termo no glossario",
        "description": "Cria um novo termo. O slug e gerado automaticamente. relatedSlugs aceita array de slugs de termos relacionados.",
        "operationId": "createGlossaryTerm",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "term",
                  "definition"
                ],
                "properties": {
                  "term": {
                    "type": "string",
                    "example": "API Key"
                  },
                  "definition": {
                    "type": "string",
                    "example": "Chave de autenticacao unica que identifica e autoriza aplicacoes a acessar uma API."
                  },
                  "content": {
                    "type": "string",
                    "description": "Conteudo HTML expandido (opcional)",
                    "example": "<h2>O que e uma API Key?</h2><p>Uma API Key (chave de API) e um identificador unico...</p><h2>Tipos de API Key</h2><p>Existem dois tipos principais...</p>"
                  },
                  "seoTitle": {
                    "type": "string",
                    "example": "API Key - O que e e como usar | Glossario Zé da API"
                  },
                  "seoDescription": {
                    "type": "string",
                    "example": "Entenda o que e uma API Key, como funciona e como usar na Zé da API."
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published"
                    ],
                    "default": "draft",
                    "example": "published"
                  },
                  "relatedSlugs": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Slugs de termos relacionados no glossario",
                    "example": [
                      "api",
                      "webhook",
                      "autenticacao"
                    ]
                  }
                }
              },
              "examples": {
                "termo-simples": {
                  "summary": "Termo simples (so definicao)",
                  "value": {
                    "term": "QR Code",
                    "definition": "Codigo bidimensional que pode ser escaneado para conectar rapidamente um dispositivo ao WhatsApp.",
                    "status": "published",
                    "relatedSlugs": [
                      "whatsapp-business-api",
                      "instancia"
                    ]
                  }
                },
                "termo-completo": {
                  "summary": "Termo com conteudo expandido e SEO",
                  "value": {
                    "term": "API Key",
                    "definition": "Chave de autenticacao unica que identifica e autoriza aplicacoes a acessar uma API.",
                    "content": "<h2>O que e uma API Key?</h2><p>Uma API Key e um identificador unico usado para autenticar requisicoes a uma API. Ela funciona como uma senha que identifica sua aplicacao.</p><h2>Boas praticas</h2><ul><li>Nunca compartilhe sua API Key publicamente</li><li>Use variaveis de ambiente para armazenar</li><li>Rotacione as chaves periodicamente</li></ul>",
                    "seoTitle": "API Key - O que e e como usar | Glossario Zé da API",
                    "seoDescription": "Entenda o que e uma API Key, como funciona e as melhores praticas de seguranca.",
                    "status": "published",
                    "relatedSlugs": [
                      "api",
                      "webhook",
                      "autenticacao"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Termo criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlossaryTerm"
                },
                "example": {
                  "id": "cm5gls003",
                  "term": "API Key",
                  "slug": "api-key",
                  "definition": "Chave de autenticacao unica que identifica e autoriza aplicacoes a acessar uma API.",
                  "content": "<h2>O que e uma API Key?</h2><p>...</p>",
                  "status": "published",
                  "relatedSlugs": [
                    "api",
                    "webhook",
                    "autenticacao"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Campos obrigatorios faltando",
            "content": {
              "application/json": {
                "example": {
                  "error": "Missing required fields: term, definition"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/admin/content/glossary/terms/{id}": {
      "get": {
        "tags": [
          "Glossary Terms"
        ],
        "summary": "Obter termo do glossario por ID",
        "operationId": "getGlossaryTerm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5gls001"
          }
        ],
        "responses": {
          "200": {
            "description": "Termo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlossaryTerm"
                },
                "example": {
                  "id": "cm5gls001",
                  "term": "Webhook",
                  "slug": "webhook",
                  "definition": "Mecanismo de notificacao HTTP que envia dados em tempo real quando um evento ocorre.",
                  "content": "<p>Webhooks sao callbacks HTTP que permitem que aplicacoes recebam notificacoes em tempo real. Quando um evento ocorre (por exemplo, uma mensagem WhatsApp recebida), o servidor envia automaticamente um POST HTTP para a URL configurada.</p><h2>Como funciona</h2><ol><li>Voce configura uma URL de callback</li><li>Um evento ocorre no WhatsApp</li><li>A Ze da API envia um POST com os dados do evento</li><li>Seu servidor processa e responde com 200 OK</li></ol>",
                  "status": "published",
                  "seoTitle": "Webhook - O que e | Glossario Ze da API",
                  "seoDescription": "Entenda o que sao webhooks e como usa-los na Ze da API.",
                  "relatedSlugs": [
                    "api",
                    "http",
                    "callback"
                  ],
                  "createdAt": "2026-01-15T10:00:00.000Z",
                  "updatedAt": "2026-02-10T14:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Glossary Terms"
        ],
        "summary": "Atualizar termo do glossario",
        "description": "Atualiza parcialmente um termo. Se o nome mudar, o slug e recalculado.",
        "operationId": "updateGlossaryTerm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5gls001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "term": {
                    "type": "string"
                  },
                  "definition": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "seoTitle": {
                    "type": "string"
                  },
                  "seoDescription": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published"
                    ]
                  },
                  "relatedSlugs": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "examples": {
                "publicar": {
                  "summary": "Publicar termo rascunho",
                  "value": {
                    "status": "published"
                  }
                },
                "atualizar-definicao": {
                  "summary": "Atualizar definicao e termos relacionados",
                  "value": {
                    "definition": "Nova definicao mais clara e completa do termo.",
                    "relatedSlugs": [
                      "api",
                      "http",
                      "callback",
                      "endpoint"
                    ]
                  }
                },
                "adicionar-conteudo": {
                  "summary": "Adicionar conteudo expandido",
                  "value": {
                    "content": "<h2>Explicacao detalhada</h2><p>Este termo refere-se a...</p>",
                    "seoTitle": "Webhook - Definicao Completa | Glossario",
                    "seoDescription": "Entenda o que sao webhooks e como usa-los."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Termo atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlossaryTerm"
                },
                "example": {
                  "id": "cm5gls001",
                  "term": "Webhook",
                  "slug": "webhook",
                  "definition": "Nova definicao mais clara e completa do termo.",
                  "content": "<h2>Explicacao detalhada</h2><p>Este termo refere-se a um mecanismo de callback HTTP...</p>",
                  "status": "published",
                  "seoTitle": "Webhook - Definicao Completa | Glossario",
                  "seoDescription": "Entenda o que sao webhooks e como usa-los.",
                  "relatedSlugs": [
                    "api",
                    "http",
                    "callback",
                    "endpoint"
                  ],
                  "createdAt": "2026-01-15T10:00:00.000Z",
                  "updatedAt": "2026-02-21T17:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "Glossary Terms"
        ],
        "summary": "Excluir termo do glossario",
        "operationId": "deleteGlossaryTerm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5gls001"
          }
        ],
        "responses": {
          "200": {
            "description": "Termo excluido",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}